code/requirements.txt
streamlit
streamlit_chat
langchain
sentence_transformers
pinecone-client
openai
json


